+++ src/components/entities/platformsFilters/SearchFiltersList/SearchFiltersList.module.css
@@ -1,0 +1,4 @@
+.filtersListWrapper {
+    display: flex;
+    flex-direction: column;
+}
